endpoints

+++ src/main/java/com/proyecto/proyecto/DTO/TratamientoDTO.java
@@ -11,6 +11,6 @@ public class TratamientoDTO {
     private String nombre;
     private double precio;
     private String descripcion;
-    private double duracion;
+    private int duracion;
     private List<ProfesionalDatosDTO> profesionales;
 }
